StreamingClient
* Errors handling
